refactor: reorganiza pacotes e renomeia repositório
- Move interfaces de `usecases` para `ports.api` para melhor separação de responsabilidades.
- Renomeia `UserJpaRepository` para `UserRepository`, mantendo consistência com as práticas do projeto.

Refs: GH-1
