+++ src/main/java/dev/ghonda/project/management/users/ports/api/usecases/UpdateUserUseCase.java
@@ -1,4 +1,4 @@
-package dev.ghonda.project.management.users.usecases;
+package dev.ghonda.project.management.users.ports.api.usecases;
 
 import dev.ghonda.project.management.users.rest.dto.UpdateUserPayload;
 import dev.ghonda.project.management.users.rest.dto.UserDetailPayload;
