+++ src/main/java/dev/ghonda/project/management/users/rest/UserController.java
@@ -6,11 +6,11 @@ import dev.ghonda.project.management.shared.dto.Resource;
 import dev.ghonda.project.management.users.rest.dto.RegisterUserPayload;
 import dev.ghonda.project.management.users.rest.dto.UpdateUserPayload;
 import dev.ghonda.project.management.users.rest.dto.UserDetailPayload;
-import dev.ghonda.project.management.users.usecases.DeleteUserUseCase;
-import dev.ghonda.project.management.users.usecases.RegisterUserUseCase;
-import dev.ghonda.project.management.users.usecases.SearchUserByIdUseCase;
-import dev.ghonda.project.management.users.usecases.SearchUsersUseCase;
-import dev.ghonda.project.management.users.usecases.UpdateUserUseCase;
+import dev.ghonda.project.management.users.ports.api.usecases.DeleteUserUseCase;
+import dev.ghonda.project.management.users.ports.api.usecases.RegisterUserUseCase;
+import dev.ghonda.project.management.users.ports.api.usecases.SearchUserByIdUseCase;
+import dev.ghonda.project.management.users.ports.api.usecases.SearchUsersUseCase;
+import dev.ghonda.project.management.users.ports.api.usecases.UpdateUserUseCase;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
