+++ src/main/java/dev/ghonda/project/management/users/usecases/impl/DeleteUserUseCaseImpl.java
@@ -1,8 +1,8 @@
 package dev.ghonda.project.management.users.usecases.impl;
 
 import dev.ghonda.project.management.shared.annotations.UseCase;
-import dev.ghonda.project.management.users.domain.UserJpaRepository;
-import dev.ghonda.project.management.users.usecases.DeleteUserUseCase;
+import dev.ghonda.project.management.users.ports.api.repositories.UserRepository;
+import dev.ghonda.project.management.users.ports.api.usecases.DeleteUserUseCase;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.transaction.annotation.Transactional;
@@ -12,7 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 @AllArgsConstructor
 public class DeleteUserUseCaseImpl implements DeleteUserUseCase {
 
-    private final UserJpaRepository userJpaRepository;
+    private final UserRepository userRepository;
 
     @Override
     @Transactional
@@ -20,7 +20,7 @@ public class DeleteUserUseCaseImpl implements DeleteUserUseCase {
         if (log.isInfoEnabled()) { log.info("Iniciando exclusão do usuário. userId: {}", userId); }
         if (log.isDebugEnabled()) log.debug("m=execute(userId: {})", userId);
 
-        this.userJpaRepository.deleteById(userId);
+        this.userRepository.deleteById(userId);
 
         if (log.isInfoEnabled()) { log.info("Usuário excluído com sucesso. userId: {}", userId); }
     }
