+++ src/main/java/dev/ghonda/project/management/users/usecases/impl/SearchUserByIdUseCaseImpl.java
@@ -2,9 +2,9 @@ package dev.ghonda.project.management.users.usecases.impl;
 
 import dev.ghonda.project.management.shared.annotations.UseCase;
 import dev.ghonda.project.management.shared.exceptions.ResourceNotFoundException;
-import dev.ghonda.project.management.users.domain.UserJpaRepository;
+import dev.ghonda.project.management.users.ports.api.repositories.UserRepository;
 import dev.ghonda.project.management.users.rest.dto.UserDetailPayload;
-import dev.ghonda.project.management.users.usecases.SearchUserByIdUseCase;
+import dev.ghonda.project.management.users.ports.api.usecases.SearchUserByIdUseCase;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -13,14 +13,14 @@ import lombok.extern.slf4j.Slf4j;
 @AllArgsConstructor
 public class SearchUserByIdUseCaseImpl implements SearchUserByIdUseCase {
 
-    private final UserJpaRepository userJpaRepository;
+    private final UserRepository userRepository;
 
     @Override
     public UserDetailPayload execute(final Long userId) {
         if (log.isInfoEnabled()) log.info("Buscando usuário pelo id: {}", userId);
         if (log.isDebugEnabled()) log.debug("m=execute(userId: {})", userId);
 
-        return this.userJpaRepository.findById(userId)
+        return this.userRepository.findById(userId)
             .map(UserDetailPayload::of)
             .orElseThrow(() -> new ResourceNotFoundException("User not found"));
     }
