+++ src/main/java/dev/ghonda/project/management/users/usecases/impl/SearchUsersUseCaseImpl.java
@@ -1,9 +1,9 @@
 package dev.ghonda.project.management.users.usecases.impl;
 
 import dev.ghonda.project.management.shared.annotations.UseCase;
-import dev.ghonda.project.management.users.domain.UserJpaRepository;
+import dev.ghonda.project.management.users.ports.api.repositories.UserRepository;
 import dev.ghonda.project.management.users.rest.dto.UserDetailPayload;
-import dev.ghonda.project.management.users.usecases.SearchUsersUseCase;
+import dev.ghonda.project.management.users.ports.api.usecases.SearchUsersUseCase;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -14,7 +14,7 @@ import org.springframework.data.domain.Pageable;
 @AllArgsConstructor
 public class SearchUsersUseCaseImpl implements SearchUsersUseCase {
 
-    private final UserJpaRepository userJpaRepository;
+    private final UserRepository userRepository;
 
     @Override
     public Page<UserDetailPayload> execute(
@@ -24,7 +24,7 @@ public class SearchUsersUseCaseImpl implements SearchUsersUseCase {
         if (log.isInfoEnabled()) { log.info("Procurando usuários com o termo de busca: {}", searchTerm); }
         if (log.isDebugEnabled()) log.debug("m=execute(searchTerm: {}, pageable: {})", searchTerm, pageable);
 
-        return this.userJpaRepository.findAll(searchTerm, pageable)
+        return this.userRepository.findAll(searchTerm, pageable)
             .map(UserDetailPayload::of);
     }
