+++ src/main/java/dev/ghonda/project/management/users/usecases/impl/UpdateUserUseCaseImpl.java
@@ -4,10 +4,10 @@ import dev.ghonda.project.management.shared.annotations.UseCase;
 import dev.ghonda.project.management.shared.exceptions.ResourceNotFoundException;
 import dev.ghonda.project.management.shared.validators.ValidatorService;
 import dev.ghonda.project.management.users.domain.Role;
-import dev.ghonda.project.management.users.domain.UserJpaRepository;
+import dev.ghonda.project.management.users.ports.api.repositories.UserRepository;
 import dev.ghonda.project.management.users.rest.dto.UpdateUserPayload;
 import dev.ghonda.project.management.users.rest.dto.UserDetailPayload;
-import dev.ghonda.project.management.users.usecases.UpdateUserUseCase;
+import dev.ghonda.project.management.users.ports.api.usecases.UpdateUserUseCase;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,7 +19,7 @@ import java.util.Optional;
 @AllArgsConstructor
 public class UpdateUserUseCaseImpl implements UpdateUserUseCase {
 
-    private final UserJpaRepository userJpaRepository;
+    private final UserRepository userRepository;
 
     private final ValidatorService validatorService;
 
@@ -31,7 +31,7 @@ public class UpdateUserUseCaseImpl implements UpdateUserUseCase {
             log.debug("m=execute(userId: {}, payload: {})", userId, payload);
         }
 
-        final var user = this.userJpaRepository.findById(userId)
+        final var user = this.userRepository.findById(userId)
             .orElseThrow(() -> new ResourceNotFoundException("Usuário não encontrado"));
 
         Optional.ofNullable(payload.username())
@@ -46,7 +46,7 @@ public class UpdateUserUseCaseImpl implements UpdateUserUseCase {
 
         this.validatorService.validate(user);
 
-        this.userJpaRepository.save(user);
+        this.userRepository.save(user);
 
         if (log.isInfoEnabled()) { log.info("Dados do usuário atualizados com sucesso. userId: {}", userId); }
         return UserDetailPayload.of(user);
